Compilation: kill warnings

+++ src/complete-sheet.c
@@ -13,6 +13,7 @@
 #include <gnumeric-config.h>
 #include "gnumeric.h"
 #include "complete-sheet.h"
+#include "gnumeric-conf.h"
 
 #include "sheet.h"
 #include "cell.h"
@@ -85,7 +86,7 @@ complete_sheet_search_iteration (Complete *complete)
 	CompleteSheet *cs = COMPLETE_SHEET (complete);
 	int i;
 
-	if (strlen (complete->text) < 
+	if ((int)strlen (complete->text) < 
 	    gnm_conf_get_core_gui_editing_autocomplete_min_chars ())
 		return FALSE;
